Added sample validation logic
I have narrowed down the list of relevant request statuses in
extension.json, otherwise the extension would try to react to unneeded
requests.

The business logic has been implemented in extension.py. The logic
checks for the quantity of items ordered, makes a decision and fails or
approves the request.